requirements.txt
Django==5.1.3
line-bot-sdk==3.14.2
python-decouple==3.8
mysqlclient==2.2.0 